+++ src/components/Footer/Footer.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {MainContainer, Imagem} from './styled';
 import { useNavigate, useLocation } from 'react-router-dom';
-import { goToHome, goToCart, goToProfile} from '../../routes/Cordinator';
+import { goToFeed, goToCart, goToProfile} from '../../routes/Cordinator';
 import home from '../../assets/img/footer/homepage.png'
 import homeColor from '../../assets/img/footer/homepage-color.png'
 import shopping from '../../assets/img/footer/shopping-cart.png'
@@ -19,14 +19,14 @@ export const Footer = () => {
     <div>
       {location.pathname === '/' ?
         <MainContainer>
-          <Imagem onClick={() => goToHome(navigate)} src={homeColor} alt="home" />
+          <Imagem onClick={() => goToFeed(navigate)} src={homeColor} alt="home" />
           <Imagem onClick={() => goToCart(navigate)} src={shopping} alt="carrinho" />
           <Imagem onClick={() => goToProfile(navigate)} src={perfil} alt="perfil" />
         </MainContainer> : '' } 
 
       {location.pathname === '/cart' ?
         <MainContainer>
-          <Imagem onClick={() => goToHome(navigate)} src={home} alt="home" />
+          <Imagem onClick={() => goToFeed(navigate)} src={home} alt="home" />
           <Imagem onClick={() => goToCart(navigate)} src={shoppingColor} alt="carrinho" />
           <Imagem onClick={() => goToProfile(navigate)} src={perfil} alt="perfil" />
         </MainContainer> : '' } 
@@ -34,7 +34,7 @@ export const Footer = () => {
 
       {location.pathname === '/profile' ?
         <MainContainer>
-          <Imagem onClick={() => goToHome(navigate)} src={home} alt="home" />
+          <Imagem onClick={() => goToFeed(navigate)} src={home} alt="home" />
           <Imagem onClick={() => goToCart(navigate)} src={shopping} alt="carrinho" />
           <Imagem onClick={() => goToProfile(navigate)} src={perfilColor} alt="perfil" />
         </MainContainer> : '' }
