+++ src/pages/SingUpPage/SingUpPage.js
@@ -3,7 +3,7 @@ import SignUpForm from './SignUpForm'
 import { ButtonBack, FormContainer, Hr, IMG, IMGBT, LogoContainer, Span } from './styled'
 import logo from '../../assets/img/logo.png'
 import back from '../../assets/img/back.png'
-import { goToLogin } from '../../routC'
+import { goToLogin } from '../../routes/Cordinator'
 import { useNavigate } from 'react-router-dom'
 
 export const SingUpPage = () => {
