--- src/routC.js
@@ -1,13 +1,0 @@
-
-
-export const goToSignUp = (navigate) => {
-    navigate('/signup')
-}
-
-export const goToEditSignUp = (navigate) => {
-    navigate('/signup-edit')
-}
-
-export const goToLogin = (navigate) => {
-    navigate('/')
-}
