+++ src/routes/Cordinator.js
@@ -1,5 +1,9 @@
-export const goToHome = (navigate) => {
-  navigate('/');
+export const goToLogin = (navigate) => {
+  navigate('/')
+}
+
+export const goToFeed = (navigate) => {
+  navigate('/feed');
 };
 
 export const goToCart = (navigate) => {
@@ -9,3 +13,11 @@ export const goToCart = (navigate) => {
 export const goToProfile = (navigate) => {
   navigate('/profile');
 };
+
+export const goToSignUp = (navigate) => {
+  navigate('/signup')
+}
+
+export const goToEditSignUp = (navigate) => {
+  navigate('/signup-edit')
+}
